fix: remove service interface
